Update AutoRaid.au3
Make the resource (elix, gold, dark) minimums more or less flexible.  By having a high scoreFactor the functionality will not change from previous version.  With a low scoreFactor then one resource higher than the minimum will offset a missed minimum on another resource.  Dark has a scaling factor (darkMultiplier), since it should be weighted in the scoring.

I hope you agree with the change.  I was cautious in coding and hope it is extended as a parameter into the globals.au3 or beyond.